requirements.txt
# Web scraping
playwright==1.49.1
beautifulsoup4==4.12.3
httpx==0.28.1
lxml==5.3.0

# AI / Anthropic
anthropic==0.42.0

# Database
sqlalchemy==2.0.36

# Document generation
python-docx==1.1.2
pdfkit==1.0.0
jinja2==3.1.5

# Configuration & CLI
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
click==8.1.8
rich==13.9.4

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
